college-memory-search/requirements.txt
chromadb>=0.5.0
groq>=0.4.0
python-dotenv>=1.0.0